\name{Compositionalridge-package}
\alias{Compositionalridge-package}
\docType{package}
\title{
Ridge Regression for Compositional Data
}

\description{
A few ridge regression models for compositional data with zero values present. 
These are the zero adjusted Dirichlet regression, the multivariate zero inflated logistic normal, 
the Kullback-Leibler divergence and the alpha-regression. Some relevant papers 
include Tsagris and Stewart (2018) <doi:10.1134/S1995080218030198>, Tsagris et al. (2026) <doi:10.48550/arXiv.2510.12663>, 
and Li et al. (2018) <doi:10.1007/s12561-018-9219-2>.  
}

\details{
\tabular{ll}{
Package: \tab Compositionalridge\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2026-09-13 \cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>.
}

\references{
Li Z., Lee K., Karagas M. R., Madan J. C., Hoen A. G., O'Malley A. J. and Li H. (2018).
Conditional regression based on a multivariate zero-inflated logistic-normal model for
microbiome relative abundance data.
Statistics in Biosciences, 10(3): 587--608.
}





